requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
cffi==1.15.0
charset-normalizer==2.0.12
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
nextcord==2.0.0b3
pycparser==2.21
PyNaCl==1.4.0
python-dotenv==0.20.0
six==1.16.0
yarl==1.7.2
